Split widget: timeline list
